requirements.txt
sentence-transformers~=2.2.2
apache-airflow
pika~=1.3.2
pandas~=1.5.0
nltk~=3.8.1
gensim~=4.3.1
scipy~=1.8.1
numpy~=1.22.0
